requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
backports.tarfile==1.2.0
blosc2==2.0.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.1.0
cycler==0.11.0
debugpy==1.8.9
decorator==5.1.1
deeprobust==0.2.9
docopt==0.6.2
docutils==0.20.1
entropy==0.1.5
et-xmlfile==1.1.0
executing==2.1.0
fasteners==0.18
filelock==3.12.2
flake8==6.0.0
flatbuffers==23.5.26
fonttools==4.41.1
frozenlist==1.4.1
fsspec==2024.3.1
h5py==3.11.0
hpelm==1.0.10
idna==3.4
imageio==2.34.0
importlib-resources==6.0.0
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.12.3
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.2
Jinja2==3.1.2
joblib==1.3.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.2.1
kiwisolver==1.4.4
lazy_loader==0.4
libclang==16.0.6
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mne==1.6.1
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.5
nbimporter==0.3.4
nest-asyncio==1.6.0
networkx==3.1
nh3==0.2.17
nir==1.0.1
nirtorch==1.0
nltk==3.8.1
nose==1.3.7
numba==0.58.1
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==24.2
pandas==2.0.3
parso==0.8.4
patsy==1.0.1
pickleshare==0.7.5
Pillow==10.0.0
pkginfo==1.11.1
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.48
protobuf==4.23.4
psutil==6.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.5.0
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.18.0
pynndescent==0.5.12
pyparsing==3.0.9
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.4.1
pywin32==308
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==26.2.0
readme_renderer==43.0
regex==2024.4.16
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1